update type of numthreads openmp supports ints as type for number of threads. This allows for more than 255 threads in some potential future use case.

+++ src/AwFmIndex.h
@@ -349,7 +349,7 @@ void awFmDeallocKmerSearchList(struct AwFmKmerSearchList *_RESTRICT_ const searc
 *					is stored on file, not in memory)
  */
 enum AwFmReturnCode awFmParallelSearchLocate(const struct AwFmIndex *_RESTRICT_ const index,
-		struct AwFmKmerSearchList *_RESTRICT_ const searchList, uint8_t numThreads);
+		struct AwFmKmerSearchList *_RESTRICT_ const searchList, uint32_t numThreads);
 
 
 /*
@@ -380,7 +380,7 @@ enum AwFmReturnCode awFmParallelSearchLocate(const struct AwFmIndex *_RESTRICT_ 
  *                    will likely vary from system to system. Suggested default value is 4
  */
 void awFmParallelSearchCount(const struct AwFmIndex *_RESTRICT_ const index,
-		struct AwFmKmerSearchList *_RESTRICT_ const searchList, uint8_t numThreads);
+		struct AwFmKmerSearchList *_RESTRICT_ const searchList, uint32_t numThreads);
 
 
 /*
